server/requirements.txt
tinytuya
pyserial

